% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_agreement.R
\name{getNNlist}
\alias{getNNlist}
\title{Get list of NN for each cell from NN graph in seurat object}
\usage{
getNNlist(obj, is.seurat = TRUE)
}
\arguments{
\item{obj}{Seurat object (with graph) or neighborhood graph}

\item{is.seurat}{logical indicating whether `obj` is a Seurat object (default: TRUE)}
}
\value{
list of NN for each cell
}
\description{
Get list of NN for each cell from NN graph in seurat object
}
