% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateBenchmark.R, R/utils.R
\name{getPredictedLabels}
\alias{getPredictedLabels}
\title{Get cell type prediction from benchmark output}
\usage{
getPredictedLabels(seu.int, int.name, id.col = "predicted.id",
  score.col = "score", filter_score = 0)

getPredictedLabels(seu.int, int.name, id.col = "predicted.id",
  score.col = "score", filter_score = 0)
}
\arguments{
\item{seu.int}{list of Seurat objects of label transfer output}

\item{int.name}{label for integration method}

\item{id.col}{column of predicted labels in obj metadata (default: "predicted.id")}

\item{score.col}{column of prediction score in obj metadata (default: "score")}

\item{filter_score}{numeric indicating the minimum prediction score to keep assigned label (default: 0)}

\item{seu.int}{list of Seurat objects of label transfer output}

\item{int.name}{label for integration method}

\item{id.col}{column of predicted labels in obj metadata (default: "predicted.id")}

\item{score.col}{column of prediction score in obj metadata (default: "score")}

\item{filter_score}{numeric indicating the minimum prediction score to keep assigned label (default: 0)}
}
\value{
returns data.frame of label predictions and prediction scores

returns data.frame of label predictions and prediction scores
}
\description{
Get cell type prediction from benchmark output

Get cell type prediction from benchmark output
}
