% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateBenchmark.R
\name{integrate_seuratCCA}
\alias{integrate_seuratCCA}
\title{Seurat CCA integration}
\usage{
integrate_seuratCCA(sce.list, integrate_features, reference = "RNA",
  query = "ATAC")
}
\arguments{
\item{sce.list}{list of SingleCellExperiment objects for RNA and ATAC seq datasets}

\item{integrate_features}{selected features to perform CCA on}

\item{reference}{reference dataset for FindTransferAnchors}

\item{query}{query dataset for FindTransferAnchors}
}
\value{
list of integration output (see details)
}
\description{
Seurat CCA integration
}
\details{
Function outputs list of
1) MultiAssayExperiment object containing input dataset + integrated (imputed) dataset for RNA,
2) misc= CCA model (Anchor object from Seurat)
}
