% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_agreement.R
\name{KNN_purity_test}
\alias{KNN_purity_test}
\title{KS-test for true KNN ECDF against ECDF of null distribution}
\usage{
KNN_purity_test(nn.list, pred.labels, n_perm = 100,
  n_cores = detectCores())
}
\arguments{
\item{nn.list}{list of nearest neighbors}

\item{pred.labels}{data.frame of predicted labels}

\item{n_perm}{number of permutations}
}
\description{
KS-test for true KNN ECDF against ECDF of null distribution
}
