% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_agreement.R
\name{KNN_score}
\alias{KNN_score}
\title{Calculate KNN agreement score with NN list}
\usage{
KNN_score(nn.list, pred.labels)
}
\arguments{
\item{nn.list}{named list of k nearest neighbors for each cell}

\item{pred.labels}{named vector of predicted labels per cell}
}
\value{
vector of KNN agreement score for each cell
}
\description{
Calculate KNN agreement score with NN list
}
