% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateBenchmark.R
\name{labelTransfer_conos}
\alias{labelTransfer_conos}
\title{Label transfer from Conos}
\usage{
labelTransfer_conos(conos.out, sce.list, annotation.col = "annotation",
  reference = "RNA", query = "ATAC")
}
\arguments{
\item{conos.out}{model output of `run_conos`}

\item{sce.list}{list of SingleCellExperiment objects used as input of `run_conos` (saved as `input` in output list of `run_conos`)}

\item{annotation.col}{column of meta}

\item{reference}{reference dataset for FindTransferAnchors}

\item{query}{query dataset for FindTransferAnchors}
}
\value{
list of Seurat objects with predicted annotations in the query dataset metadata
}
\description{
Label transfer from Conos
}
