% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateBenchmark.R
\name{labelTransfer_liger}
\alias{labelTransfer_liger}
\title{Label transfer for LIGER}
\usage{
labelTransfer_liger(liger.obj, sce.list, annotation.col = "annotation",
  k = 50, reference = "RNA", query = "ATAC")
}
\arguments{
\item{liger.obj}{liger object (output of run_liger)}

\item{sce.list}{list of SingleCellExperiment objects used as input for run_liger}

\item{annotation.col}{column of meta}

\item{reference}{reference dataset for label transfer}

\item{query}{query dataset for label transfer}
}
\value{
list of Seurat objects with predicted annotations (with score) in the query dataset metadata
}
\description{
Label transfer for LIGER
}
\details{
Label transfer is performed by computing a cross-dataset k-nearest neighbor graph in the aligned factor space.
Then we set the value of each missing label is set to the most abundant label among its k nearest neighbors in the reference dataset.
Ties are dealt with by taking the first of the maximum values. As prediction score I take the number of neighbors harboring the predicted.id
divided by k.
}
