% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_agreement.R
\name{null_KNN_score}
\alias{null_KNN_score}
\title{Calculate null KNN score for predicted labels}
\usage{
null_KNN_score(nn.list, pred.labels, n_perm = 10, n_cores = 10)
}
\description{
Calculate null KNN score for predicted labels
}
\details{
Implements a permutation test to avoid that a high KNN score is assigned to a label transfer method that just 
gives the same label to everything
}
