% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateBenchmark.R
\name{propagateNNannotation}
\alias{propagateNNannotation}
\title{Utility function for labelTransfer with Liger}
\usage{
propagateNNannotation(nn.vector, annotation)
}
\description{
Utility function for labelTransfer with Liger
}
