% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateBenchmark.R
\name{run_liger}
\alias{run_liger}
\title{LIGER NMF model}
\usage{
run_liger(sce.list, integrate_features, reference = "RNA",
  query = "ATAC")
}
\arguments{
\item{sce.list}{list of SingleCellExperiment objects for RNA and ATAC seq datasets}

\item{integrate_features}{selected features to perform CCA on}

\item{reference}{reference dataset for FindTransferAnchors}

\item{query}{query dataset for FindTransferAnchors}
}
\value{
list of integration output (see details)
}
\description{
LIGER NMF model
}
\details{
Function outputs list of
1) model:
2) misc
}
