% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateBenchmark.R
\name{run_SeuratCCA}
\alias{run_SeuratCCA}
\title{Seurat CCA integration model}
\usage{
run_SeuratCCA(sce.list, integrate_features, reference = "RNA",
  query = "ATAC")
}
\arguments{
\item{sce.list}{list of SingleCellExperiment objects for RNA and ATAC seq datasets}

\item{integrate_features}{selected features to perform CCA on}

\item{reference}{reference dataset for FindTransferAnchors}

\item{query}{query dataset for FindTransferAnchors}
}
\value{
list of integration output (see details)
}
\description{
Seurat CCA integration model
}
\details{
Function outputs list of
1) model: CCA model (Anchor object from Seurat)
2) input: Seurat object list containing input datasets
}
