% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeatures.R
\name{select_highlyCovered}
\alias{select_highlyCovered}
\title{Select highly covered features}
\usage{
select_highlyCovered(sce, frac_cells = 0.1)
}
\arguments{
\item{sce}{SingleCellExperiment object (must have counts and logcounts for log-normalized data)}

\item{frac_cells}{fraction of cells in which a feature has to be different from 0 to be included}
}
\description{
Select highly covered features
}
