% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeatures.R
\name{select_highlyVariable}
\alias{select_highlyVariable}
\title{Select highly variable features
Wrapper around Seurat function}
\usage{
select_highlyVariable(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object (must have counts and logcounts for log-normalized data)}
}
\description{
Select highly variable features
Wrapper around Seurat function
}
